third_party/blink/web_tests/external/wpt/app-history/ordering-and-transition/currentchange-before-popsate-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  await appHistory.navigate("#foo").committed;
  assert_equals(appHistory.entries().length, 2);

  appHistory.onnavigate = e => e.transitionWhile(Promise.resolve());

  let oncurrentchange_back_called = false;
  let onpopstate_back_called = false;
  window.onpopstate = t.step_func(e => {
    onpopstate_back_called = true;
    assert_true(oncurrentchange_back_called);
  });
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_back_called = true;
    assert_false(onpopstate_back_called);
  });
  let back_result = appHistory.back();
  assert_false(oncurrentchange_back_called);
  assert_false(onpopstate_back_called);
  await back_result.finished;
  assert_true(oncurrentchange_back_called);
  assert_true(onpopstate_back_called);

  let oncurrentchange_forward_called = false;
  let onpopstate_forward_called = false;
  window.onpopstate = t.step_func(e => {
    onpopstate_forward_called = true;
    assert_true(oncurrentchange_forward_called);
  });
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_forward_called = true;
    assert_false(onpopstate_forward_called);
  });
  let forward_result = appHistory.forward();
  assert_false(oncurrentchange_forward_called);
  assert_false(onpopstate_forward_called);
  await forward_result.finished;
  assert_true(oncurrentchange_back_called);
  assert_true(onpopstate_forward_called);
}, "AppHistoryCurrentChangeEvent fires for appHistory.back() and appHistory.forward()");
</script>
